Add abstract controller base class
