+++ Startup.cs
@@ -2,13 +2,12 @@
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.DependencyInjection.Extensions;
 
 namespace GitHubAppDotnetHelloWorld
 {
     public class Startup
     {
-
-        public static IConfiguration Configuration { get; private set; }
 
         // This method gets called by the runtime. Use this method to add services to the container.
         // For more information on how to configure your application, visit https://go.microsoft.com/fwlink/?LinkID=398940
@@ -17,6 +16,7 @@ namespace GitHubAppDotnetHelloWorld
             services
                 .AddMvcCore()
                 .AddGitHubWebHooks();
+
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
